pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "img_validator"
dynamic = ["version"]
description = "A Python data science project"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "python-dotenv",
    "exif"
]

[project.scripts]
img_validator = "bin:main.main"

[project.optional-dependencies]
dev = [
    "ipykernel",
    "pytest",
]
